+++ src/matching/cosmetics.ts
@@ -1,20 +1,17 @@
 import { getPublicSuffix } from 'tldts';
 
 import { CosmeticFilter } from '../parsing/cosmetic-filter';
-import { fastStartsWith } from '../utils';
 
 /* Checks that hostnamePattern matches at the end of the hostname.
  * Partial matches are allowed, but hostname should be a valid
  * subdomain of hostnamePattern.
  */
-function checkHostnamesPartialMatch(hostname: string, hostnamePattern: string): boolean {
-  let pattern = hostnamePattern;
-  if (fastStartsWith(hostnamePattern, '~')) {
-    pattern = pattern.slice(1);
-  }
-
-  if (hostname.endsWith(pattern)) {
-    const patternIndex = hostname.length - pattern.length;
+function checkHostnamesPartialMatch(
+  hostname: string,
+  hostnamePattern: string,
+): boolean {
+  if (hostname.endsWith(hostnamePattern)) {
+    const patternIndex = hostname.length - hostnamePattern.length;
     if (patternIndex === 0 || hostname[patternIndex - 1] === '.') {
       return true;
     }
@@ -29,7 +26,10 @@ function checkHostnamesPartialMatch(hostname: string, hostnamePattern: string): 
  * It takes care of the concept of entities introduced by uBlock: google.*
  * https://github.com/gorhill/uBlock/wiki/Static-filter-syntax#entity-based-cosmetic-filters
  */
-function matchHostname(hostname: string, hostnamePattern: string): boolean {
+function matchHostname(
+  hostname: string,
+  hostnamePattern: string,
+): boolean {
   if (hostnamePattern.endsWith('.*')) {
     // Match entity:
     const entity = hostnamePattern.slice(0, -2);
@@ -53,24 +53,31 @@ function matchHostname(hostname: string, hostnamePattern: string): boolean {
   return checkHostnamesPartialMatch(hostname, hostnamePattern);
 }
 
-export default function matchCosmeticFilter(
-  filter: CosmeticFilter,
-  hostname: string,
-): { hostname: string } | null {
+export default function matchCosmeticFilter(filter: CosmeticFilter, hostname: string): boolean {
   // Check hostnames
   if (filter.hasHostnames()) {
     if (hostname) {
       const hostnames = filter.getHostnames();
+
+      // Check for exceptions
       for (let i = 0; i < hostnames.length; i += 1) {
-        if (matchHostname(hostname, hostnames[i])) {
-          return { hostname: hostnames[i] };
+        const filterHostname = hostnames[i];
+        if (filterHostname[0] === '~' && matchHostname(hostname, filterHostname.slice(1))) {
+          return false;
+        }
+      }
+
+      // Check for positive matches
+      for (let i = 0; i < hostnames.length; i += 1) {
+        const filterHostname = hostnames[i];
+        if (filterHostname[0] !== '~' && matchHostname(hostname, filterHostname)) {
+          return true;
         }
       }
     }
 
-    // No hostname match
-    return null;
+    return false;
   }
 
-  return { hostname: '' };
+  return true;
 }