+++ test/parsing.test.ts
@@ -773,47 +773,6 @@ describe('Cosmetic filters', () => {
       selector: 'script.js, arg1, arg2, arg3',
     });
   });
-
-  describe('tokenizes filter', () => {
-    [
-      // Plain selectors
-      { selector: '.c', tokens: ['.c'] },
-      { selector: '.c.d', tokens: ['.c.d'] },
-      { selector: '.c .d', tokens: ['.c', '.d'] },
-
-      // With styles included (brackets)
-      { selector: '.c[foo]', tokens: ['.c'] },
-      { selector: '[foo].c', tokens: ['.c'] },
-      { selector: '[foo].c[foo]', tokens: ['.c'] },
-      { selector: '[foo[bar]].c[foo]', tokens: ['.c'] },
-      { selector: '[foo[bar]].c[foo].d', tokens: ['.c', '.d'] },
-      { selector: '[foo[bar]].c[foo[baz]].d', tokens: ['.c', '.d'] },
-      { selector: '.c[foo[bar]].d[foo[baz]].e', tokens: ['.c', '.d', '.e'] },
-
-      // With combinators
-      { selector: '.b > .c', tokens: ['.c'] },
-      { selector: '.a ~ .b > .c', tokens: ['.c'] },
-      { selector: '.a ~ .b ~ .c', tokens: ['.c'] },
-      { selector: '.a + .b ~ .c', tokens: ['.c'] },
-      { selector: '.a + .b + .c', tokens: ['.c'] },
-
-      // With combinators + styles
-      { selector: '.c[foo[bar]].d[foo[baz]].e > .c', tokens: ['.c'] },
-      { selector: '.a > .c[foo[bar]].d[foo[baz]].e ~ .c', tokens: ['.c'] },
-      { selector: '.a > .c[foo[bar]].d[foo[baz]].e ~ .c[foo]', tokens: ['.c'] },
-      { selector: '.a > .c[foo[bar]].d[foo[baz]].e ~ .c[foo[bar]].d', tokens: ['.c', '.d'] },
-    ].forEach((testCase) => {
-      it(testCase.selector, () => {
-        const parsed = parseCosmeticFilter(`##${testCase.selector}`);
-        expect(parsed).not.toBeNull();
-        if (parsed !== null) {
-          expect(parsed.getTokensSelector()).toEqual(
-            new Uint32Array(testCase.tokens.map(fastHash)),
-          );
-        }
-      });
-    });
-  });
 });
 
 describe('Filters list', () => {